Move atom components with their stories into seperate directories
