+++ .vscode/settings.json
@@ -7,35 +7,42 @@
     "workbench.iconTheme": "material-icon-theme",
 
     "material-icon-theme.activeIconPack": "react",
-    "material-icon-theme.hidesExplorerArrows": true,
-    "material-icon-theme.showUpdateMessage": true,
-    "material-icon-theme.folders.color": "#ec407a",
-    "material-icon-theme.folders.associations": {
-        ".cache": "Database",
-        "design_assets": "Theme",
-        "design_files": "Custom",
-        "node_modules": "Node",
-        "public": "Public",
-        "page-data": "Global",
-        "01_atoms": "Cluster",
-        "02_molecules": "App",
-        "03_organisms": "Components",
-        "04_templates": "Layout",
-        "05_pages": "Log",
-        "06_posts": "Docs",
-        "07_images": "Images",
-        "08_icons": "Svg",
-        "09_utilities": "Utils",
-        "10_styles": "Sass",
-        "00_base-styles": "Theme",
-        "01_atom-styles": "Cluster",
-        "02_molecule-styles": "App",
-        "03_organism-styles": "Components",
-        "04_template-styles": "Layout"
-    },
-    "material-icon-theme.files.associations": {
-        "*.js": "react"
-    },
+	"material-icon-theme.hidesExplorerArrows": true,
+	"material-icon-theme.showUpdateMessage": true,
+	"material-icon-theme.folders.color": "#ec407a",
+	"material-icon-theme.folders.associations": {
+		".cache": "Database",
+		"design_assets": "Theme",
+		"design_files": "Custom",
+		"node_modules": "Node",
+		"public": "Public",
+		"page-data": "Global",
+		"01_atoms": "Cluster",
+		"Button": "Cluster",
+		"ExLink": "Cluster",
+		"Icon": "Cluster",
+		"Image": "Cluster",
+		"InLink": "Cluster",
+		"NavLink": "Cluster",
+		"SkipLink": "Cluster",
+		"02_molecules": "App",
+		"03_organisms": "Components",
+		"04_templates": "Layout",
+		"05_pages": "Log",
+		"06_posts": "Docs",
+		"07_images": "Images",
+		"08_icons": "Svg",
+		"09_utilities": "Utils",
+		"10_styles": "Sass",
+		"00_base-styles": "Theme",
+		"01_atom-styles": "Cluster",
+		"02_molecule-styles": "App",
+		"03_organism-styles": "Components",
+		"04_template-styles": "Layout"
+	},
+	"material-icon-theme.files.associations": {
+		"*.js": "react"
+	},
 
     "scss.lint.duplicateProperties": "warning",
     "scss.lint.unknownProperties": "warning",
