+++ src/01_atoms/Icon/Icon.stories.js
