+++ src/01_atoms/Image/Image.stories.js
