+++ src/03_organisms/Header.js
@@ -7,7 +7,7 @@
 
 import React from "react"
 import { useStaticQuery, graphql } from "gatsby"
-import InLink from "../01_atoms/InLink"
+import InLink from "../01_atoms/InLink/InLink"
 import MenuMobile from "../02_molecules/MenuMobile"
 import MenuDesktop from "../02_molecules/MenuDesktop"
