+++ src/04_templates/MainLayout.js
@@ -4,7 +4,7 @@
  */
 
 import React from "react"
-import SkipLink from "../01_atoms/SkipLink"
+import SkipLink from "../01_atoms/SkipLink/SkipLink"
 import { SkipNavContent } from "@reach/skip-nav"
 import Header from "../03_organisms/Header"
 import Footer from "../03_organisms/Footer"
