+++ src/pages/style-guide.js
@@ -5,11 +5,11 @@
 
 import React from "react"
 import Layout from "../04_templates/MainLayout"
-import Image from "../01_atoms/Image"
-import Icon from "../01_atoms/Icon"
-import InLink from "../01_atoms/InLink"
-import ExLink from "../01_atoms/ExLink"
-import Button from "../01_atoms/Button"
+import Image from "../01_atoms/Image/Image"
+import Icon from "../01_atoms/Icon/Icon"
+import InLink from "../01_atoms/InLink/InLink"
+import ExLink from "../01_atoms/ExLink/ExLink"
+import Button from "../01_atoms/Button/Button"
 
 // @todo: Move to style guide styling unless adding Storybook!?
 const Container = {
